feat: show today date in `Time` area / `Time` エリアに今日の日付を表示
- fix: ディレクトリ構成ちょっと変だったので修正
