--- App.xaml.cs
@@ -1,14 +1,0 @@
-﻿using System.Configuration;
-using System.Data;
-using System.Windows;
-
-namespace WpfApp1
-{
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application
-    {
-    }
-
-}
