--- MainWindow.xaml
@@ -1,32 +1,0 @@
-﻿<Window x:Class="WpfApp1.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="506" Width="703">
-
-    <DockPanel>
-        <Menu DockPanel.Dock="Top">
-            <MenuItem Header="ファイル">
-                <MenuItem Header="新規作成"></MenuItem>
-            </MenuItem>
-        </Menu>
-        <Grid RenderTransformOrigin="0.601,0.473">
-            <Border BorderBrush="Black" BorderThickness="1" Height="72" Margin="10,34,240,0" VerticalAlignment="Top">
-                <Label Content="作品タイトル" Height="64" VerticalAlignment="Center" VerticalContentAlignment="Center" HorizontalAlignment="Center" Width="453" FontSize="24" FontWeight="Bold" FontFamily="Meiryo UI" Padding="15,5,15,5"/>
-            </Border>
-            <Label Content="Title" HorizontalAlignment="Left" Margin="37,15,0,0" VerticalAlignment="Top" Height="27" Width="46" FontSize="24" Background="White" Padding="0,0,0,0"/>
-            <Border BorderBrush="Black" BorderThickness="1">
-                <Border BorderBrush="Black" BorderThickness="1" Height="180" Margin="9,130,240,0" VerticalAlignment="Top">
-                    <Label Content="Label" Margin="0,18,0,31" FontSize="16"/>
-                </Border>
-            </Border>
-            <Label Content="Summary" Margin="37,111,0,0" FontSize="24" RenderTransformOrigin="0.5,0.5" Padding="0,0,0,0" Background="White" ScrollViewer.CanContentScroll="True" Height="33" VerticalAlignment="Top" HorizontalAlignment="Left" Width="101"/>
-            <Border BorderBrush="Black" BorderThickness="1" Margin="10,339,467,10">
-                <TextBlock Text="{Binding Date, ConverterCulture=ja-JP}" Margin="0,9,0,0"/>
-            </Border>
-            <Label Content="Time" HorizontalAlignment="Left" Margin="37,320,0,0" VerticalAlignment="Top" Padding="0,0,0,0" FontSize="24" Background="White"/>
-        </Grid>
-    </DockPanel>
-</Window>
