--- Time.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Tips_StringFormat.Times
-{
-    public class Time
-    {
-        public Time() 
-        {
-            this.Date = DateTime.Today;
-        }
-        public DateTime Date { get; set; }
-    }
-}
