+++ WpfApp1.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35004.147
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApp1", "WpfApp1\WpfApp1.csproj", "{A32CFE80-D313-4C0A-BAD6-448E1CDBF81E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WpfApp1", "WpfApp1\WpfApp1.csproj", "{3ADE02F5-7F76-416D-8557-94D416A4ADFC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A32CFE80-D313-4C0A-BAD6-448E1CDBF81E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A32CFE80-D313-4C0A-BAD6-448E1CDBF81E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A32CFE80-D313-4C0A-BAD6-448E1CDBF81E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A32CFE80-D313-4C0A-BAD6-448E1CDBF81E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3ADE02F5-7F76-416D-8557-94D416A4ADFC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3ADE02F5-7F76-416D-8557-94D416A4ADFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3ADE02F5-7F76-416D-8557-94D416A4ADFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3ADE02F5-7F76-416D-8557-94D416A4ADFC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
